refactor: extract subscription management out of BandLink into its own module (INT-1319)
